requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
configparser==5.0.2
crawlertool==0.1.28
crayons==0.4.0
greenlet==1.1.1
idna==3.2
lxml==4.6.3
mysql-connector==2.2.9
mysql-connector-python==8.0.26
mysqlclient==2.0.3
numpy==1.21.2
pandas==1.3.2
powerline-status==2.7
protobuf==3.15.6
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
ronglian-sms-sdk==1.0.0
schedule==1.1.0
selenium==3.141.0
Selenium4R==0.1.6
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.23
webdriver-manager==3.4.2
oss2~=2.15.0
PyYAML~=5.4.1